feat(front): turn on hardware back button handler
