+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "MobileGallery",
   "private": true,
-  "version": "0.0.3",
+  "version": "0.0.4",
   "type": "module",
   "scripts": {
     "dev": "vite",
