+++ src/main.ts
@@ -13,12 +13,12 @@ import '@ionic/vue/css/structure.css'
 import '@ionic/vue/css/typography.css'
 
 /* Optional CSS utils that can be commented out */
-import '@ionic/vue/css/padding.css'
+import '@ionic/vue/css/display.css'
+import '@ionic/vue/css/flex-utils.css'
 import '@ionic/vue/css/float-elements.css'
+import '@ionic/vue/css/padding.css'
 import '@ionic/vue/css/text-alignment.css'
 import '@ionic/vue/css/text-transformation.css'
-import '@ionic/vue/css/flex-utils.css'
-import '@ionic/vue/css/display.css'
 
 /**
  * Ionic Dark Mode
@@ -41,7 +41,11 @@ import './main.css'
 import { register } from 'swiper/element/bundle'
 register()
 
-const app = createApp(App).use(IonicVue).use(router)
+const app = createApp(App)
+  .use(IonicVue, {
+    experimentalCloseWatcher: true,
+  })
+  .use(router)
 
 router.isReady().then(() => {
   app.mount('#app')
