+++ src/types/componentTypes.ts
@@ -4,7 +4,3 @@ import PreviewComponent from '@/components/media/upload/PreviewComponent.vue'
 export interface PreviewComponentRef extends InstanceType<typeof PreviewComponent> {
   emptyMedia: () => void
 }
-
-export interface ViewerComponentRef extends InstanceType<typeof ViewerComponent> {
-  getCurrentSlide: () => number
-}
