+++ src/views/auth/ExtraPage.vue
@@ -14,6 +14,18 @@
 </template>
 
 <script setup lang="ts">
+/* Import */
+import { handleBackButton } from '@/utils/functions'
 import { IonIcon, IonLabel, IonPage, IonRouterOutlet, IonTabBar, IonTabButton, IonTabs } from '@ionic/vue'
 import { qrCodeOutline } from 'ionicons/icons'
+import { onMounted } from 'vue'
+import { useRouter } from 'vue-router'
+
+/* Const */
+const router = useRouter()
+
+/* Mounted Lifecycle Hook */
+onMounted(() => {
+  handleBackButton(2, router.back)
+})
 </script>
