+++ src/views/guest/VerificationPage.vue
@@ -1,7 +1,10 @@
 <template>
-  <Verification />
+  <IonPage>
+    <Verification />
+  </IonPage>
 </template>
 
 <script setup lang="ts">
 import Verification from '@/components/authentication/VerificationPage.vue'
+import { IonPage } from '@ionic/vue'
 </script>
